namada_trans_token
Trait Keys

pub trait Keys {
    // Required methods
    fn balance_key(token: &Address, owner: &Address) -> Key;
    fn is_balance_key<'a>(
        token_addr: &Address,
        key: &'a Key,
    ) -> Option<&'a Address>;
    fn is_any_token_balance_key(key: &Key) -> Option<[&Address; 2]>;
    fn minter_key(token_addr: &Address) -> Key;
    fn parameter_prefix(token_addr: &Address) -> Key;
    fn minted_balance_key(token_addr: &Address) -> Key;
    fn is_any_minted_balance_key(key: &Key) -> Option<&Address>;
}

Abstract token keys interface

Required Methods

fn balance_key(token: &Address, owner: &Address) -> Key

Key for transparent token balance

fn is_balance_key<'a>(token_addr: &Address, key: &'a Key) -> Option<&'a Address>

Returns the owner address if the given storage key is a balance key for the given token.

fn is_any_token_balance_key(key: &Key) -> Option<[&Address; 2]>

Check if the given storage key is a balance key for an unspecified token. If it is, return the token and owner address.

fn minter_key(token_addr: &Address) -> Key

Obtain a storage key for the multitoken minter.

fn parameter_prefix(token_addr: &Address) -> Key

Obtain a storage key prefix for token parameters.

fn minted_balance_key(token_addr: &Address) -> Key

Obtain a storage key for the minted multitoken balance.

fn is_any_minted_balance_key(key: &Key) -> Option<&Address>

Check if the given storage key is for total supply of a unspecified token. If it is, returns the token.

Object Safety

This trait is not object safe.