namada_trans_token
Trait Read

pub trait Read<S> {
    // Required methods
    fn read_denom(
        storage: &S,
        token: &Address,
    ) -> Result<Option<Denomination>, Error>;
    fn get_effective_total_native_supply(storage: &S) -> Result<Amount, Error>;
    fn read_balance(
        storage: &S,
        token: &Address,
        owner: &Address,
    ) -> Result<Amount, Error>;
}

Abstract token storage read interface

Required Methods

fn read_denom( storage: &S, token: &Address, ) -> Result<Option<Denomination>, Error>

Read the denomination of a given token, if any. Note that native transparent tokens do not have this set and instead use the constant token::NATIVE_MAX_DECIMAL_PLACES.

fn get_effective_total_native_supply(storage: &S) -> Result<Amount, Error>

Get the effective circulating total supply of native tokens.

fn read_balance( storage: &S, token: &Address, owner: &Address, ) -> Result<Amount, Error>

Read the balance of a given token and owner.

Object Safety

This trait is not object safe.